+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 IMPROVEMENTS:
+ * Add support for high availability by allowing multiple instances of the autoscaler to run at the same time, but having only one actively executing.
  * build: Updated to Go 1.21.1 [[GH-707](https://github.com/hashicorp/nomad-autoscaler/pull/707)]
  * plugin/target/aws: Prevent scaling if an instance refresh is in progress [[GH-597](https://github.com/hashicorp/nomad-autoscaler/pull/597)]
  * plugin/target/aws: Add new configuration `retry_attempts` to account for potentially slow ASG update operations [[GH-594](https://github.com/hashicorp/nomad-autoscaler/pull/594)